+++ src/gdb/mipsread.c
@@ -300,7 +300,7 @@ read_alphacoff_dynamic_symtab(struct section_offsets *section_offsets,
 
       sym_value = bfd_h_get_64(abfd, (bfd_byte *)x_symp->st_value);
       sym_info = bfd_h_get_8(abfd, (bfd_byte *)x_symp->st_info);
-      sym_shndx = bfd_h_get_16(abfd, (bfd_byte *)x_symp->st_shndx);
+      sym_shndx = (uint16_t)bfd_h_get_16(abfd, (bfd_byte *)x_symp->st_shndx);
       isglobal = (ELF_ST_BIND(sym_info) == STB_GLOBAL);
 
       if (sym_shndx == SHN_UNDEF)
